CueControl: Fix cast from double to enum on MSVC
Fixes this error on MSVC:

    ..\src\engine\controls\cuecontrol.cpp(2478): error C2440: 'static_cast': cannot convert from 'double' to 'HotcueControl::Status'
    ..\src\engine\controls\cuecontrol.cpp(2478): note: Conversions between enumeration and floating point values are no longer allowed

+++ src/engine/controls/cuecontrol.cpp
@@ -2485,5 +2485,7 @@ void HotcueControl::setStatus(HotcueControl::Status status) {
 }
 
 HotcueControl::Status HotcueControl::getStatus() const {
-    return static_cast<Status>(m_pHotcueStatus->get());
+    // Cast to int before casting to the int-based enum class because MSVC will
+    // throw a hissy fit otherwise.
+    return static_cast<Status>(static_cast<int>(m_pHotcueStatus->get()));
 }
